+++ src/app/globals.css
@@ -10,6 +10,7 @@
   /* font and other theme colors */
   --color-primary: #ffC825;
   --color-secondary: #4d4d4d;
+  --color-swblue:#2B3674;
   --color-accent: #171717;
   --color-gray-shade: #a7a8aa;
   --color-primary-shade: #fffaeb;
