+++ src/app/login/layout.tsx
@@ -1,6 +1,7 @@
 import React , {ReactNode} from 'react'
 import Navbar from '../components/Navbar';
 import Authprovider from "../AuthenticatorComp/provider";
+import Footer from '../components/Footer';
 
  interface Props {
     children : ReactNode;
@@ -8,14 +9,17 @@ import Authprovider from "../AuthenticatorComp/provider";
 
  const LoginLayout = ( {children}:Props) => {
    return (
+    <>
      <div className='h-screen flex flex-col'>
-          <Authprovider>
+        <Authprovider>
           <Navbar />
         </Authprovider>
         <main className='overflow-y-auto'>
             {children}
         </main>
      </div>
+      <Footer/>
+     </>
    )
  }
