+++ src/app/signup/registration/page.tsx
@@ -1,3 +1,4 @@
+'use client'
 import React from 'react'
 import SignUpForm from './SignUpForm'
